requirements.txt
ccxt==4.2.23
pandas==2.2.3
numpy==2.1.3
ta==0.11.0
python-telegram-bot==21.6
schedule==1.2.2
requests==2.32.3
